moved OuterErrorBoundary inside blocksmith
