+++ src/plugins/blocksmith/components/blocksmith/index.jsx
@@ -1,8 +1,6 @@
 import React from 'react';
 
-// eslint-disable-next-line import/no-named-as-default
-import OuterErrorBoundary from '@triniti/cms/plugins/blocksmith/components/outer-error-boundary';
-
+import OuterErrorBoundary from './outer-error-boundary';
 import Blocksmith from './Blocksmith';
 
 export default ({ ...rest }) => (
